third_party/blink/web_tests/fast/forms/time/time-picker-select-value-step2.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../resources/common.js"></script>
<script src="../resources/picker-common.js"></script>
<script src="../calendar-picker/resources/calendar-picker-common.js"></script>
</head>
<body>
<input type="time" id="time" value="14:15:00" step="2">
<script>

let t = async_test('Test select invalid value in time popup');

function selectValue() {
  let timeElement = document.getElementById("time");
  timeElement.addEventListener("change", t.step_func_done(() => {
    assert_equals(timeElement.value, "18:16:01");
    assert_true(timeElement.validity.stepMismatch);
    assert_false(timeElement.validity.valid);
  }));
  clickTimeCellAt(0, 4);
  clickTimeCellAt(1, 1);
  clickTimeCellAt(2, 1);
  eventSender.keyDown('Enter');
}

t.step(() => {
  openPickerDeprecatedJsTest(document.getElementById('time'), t.step_func(selectValue));
});
</script>
</body>
</html>
